requirements.txt
blinker==1.9.0
boto3==1.34.11
botocore==1.34.11
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
Werkzeug==3.1.3
psycopg2-binary==2.9.9
psycopg2==2.9.9
requests==2.31.0
urllib3==2.0.7
PyJWT==2.8.0
python-jose==3.3.0
cryptography==41.0.7
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dotenv==1.0.0

# smtplib and email are part of Python standard library, no pip install needed